requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
werkzeug==2.3.7